edit files for OOP conventions.
