+++ README.md
@@ -1,4 +1,4 @@
-# Web Scraper for CHUNCHEON Bus Timetable
+# Web Scraper for ChunCheon Bus Timetable
 
 ![Version](https://img.shields.io/github/languages/code-size/altair823/ChuncheonBus)
 [![GitHub release (latest by date)](https://img.shields.io/github/v/release/altair823/ChuncheonBus)]()
