--- lib/StringExtractor.py
@@ -1,3 +1,0 @@
-# Extract meaningful Strings from html string that scraped by scraper module.
-# So this module needs the scraper module before executed.
-from . import Scraper
